Wed, 17:30, 1 file, M1 iMac
 M Alfred.alfredpreferences/workflows/user.workflow.600317EB-4C24-44F6-9700-56FB6164405D/work-mode_[Browser].applescript

+++ Alfred.alfredpreferences/workflows/user.workflow.600317EB-4C24-44F6-9700-56FB6164405D/work-mode_[Browser].applescript
@@ -1,7 +1,13 @@
 #!/usr/bin/env osascript
 
--- quit zoom without asking for confirmation
-do shell script ("killall 'zoom.us' || true")
+-- some aspects of this script depend on the device used
+set isOffice to (computer name of (system info) is "Christopher’s Mac mini")
+
+-- quit Zoom, start Spotify
+if (isOffice is false) then
+	tell application id "com.runningwithcrayons.Alfred" to run trigger "play" in workflow "com.vdesabou.spotify.mini.player"
+	do shell script ("killall 'zoom.us' || true")
+end if
 
 -- Start apps
 tell application "Mimestream" to if it is not running then activate
@@ -9,17 +15,38 @@ tell application "Brave Browser" to if it is not running then activate
 tell application "Slack" to if it is not running then activate
 tell application "Discord" to if it is not running then activate
 
-do shell script "defaults write com.iconfactory.Twitterrific5 advancedShowDockIcon -bool YES"
+-- Reset to workspace depending on location
+if (isOffice is false) then
+	set workspace to "Home"
+else
+	set workspace to "Office"
+end if
+tell application "Drafts" to open location ("drafts://x-callback-url/runaction?&action=Workspace-" & workspace)
 
--- reset Drafts (Actions contains focus mode off already)
-tell application "Drafts" to open location "drafts://x-callback-url/runaction?&action=Workspace-Home"
+-- Twitterific: no more headless
+if (isOffice is true) then
+	do shell script "defaults write com.iconfactory.Twitterrific5 advancedShowDockIcon -bool YES"
+	tell application "Twitterrific"
+		quit
+		delay 0.1
+		activate
+	end tell
+end if
 
--- jump to top in Twitterific
-tell application "Twitterrific" to if it is not running then activate
-repeat until application "Twitterrific" is running
-	delay 0.1
-end repeat
+-- Twitterific: scroll up
 tell application id "com.runningwithcrayons.Alfred" to run trigger "twitterrific-scroll-up" in workflow "de.chris-grieser.twitter-tweaks" with argument ""
 
-delay 0.2
+-- arrange windows
+delay 0.1
 tell application "Moom" to arrange windows according to snapshot "💼 Work"
+
+-- headless again
+if (isOffice is true) then
+	delay 0.1
+	do shell script "defaults write com.iconfactory.Twitterrific5 advancedShowDockIcon -bool NO"
+	tell application "Twitterrific"
+		quit
+		delay 0.1
+		activate
+	end tell
+end if
